make naming consisttant & kill dead page
