+++ packages/synapse-interface/pages/index.tsx
@@ -1,9 +1,10 @@
+import ReactGA from 'react-ga'
 import { Banner } from '@/components/Banner'
-import StateManagedBridge from './state-managed-bridge'
 import { Portfolio } from '@/components/Portfolio/Portfolio'
 import { LandingPageWrapper } from '@/components/layouts/LandingPageWrapper'
-import ReactGA from 'react-ga'
+
 import useSyncQueryParamsWithBridgeState from '@/utils/hooks/useSyncQueryParamsWithBridgeState'
+import StateManagedBridge from './bridge'
 
 // TODO: someone should add this to the .env, disable if blank, etc.
 // this is being added as a hotfix to assess user load on the synapse explorer api
